Testing update

+++ eucaops/__init__.py
@@ -34,6 +34,7 @@ from boto.ec2.image import Image
 from boto.ec2.volume import Volume
 from cwops import CWops
 from asops import ASops
+from eucaops.elbops import ELBops
 from iamops import IAMops
 from ec2ops import EC2ops
 from s3ops import S3ops
@@ -42,15 +43,19 @@ import time
 from eutester.euservice import EuserviceManager
 from boto.ec2.instance import Reservation
 from eutester.euconfig import EuConfig
+from eutester.euproperties import Euproperty_Manager
 from eutester.machine import Machine
 from eutester.euvolume import EuVolume
 from eutester import eulogger
 import re
 import os
 
-class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
+class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops, ELBops):
     
-    def __init__(self, config_file=None, password=None, keypath=None, credpath=None, aws_access_key_id=None, aws_secret_access_key = None,  account="eucalyptus", user="admin", username=None, APIVersion='2011-01-01', region=None, ec2_ip=None, s3_ip=None, as_ip=None, download_creds=True,boto_debug=0):
+    def __init__(self, config_file=None, password=None, keypath=None, credpath=None, aws_access_key_id=None,
+                 aws_secret_access_key = None,  account="eucalyptus", user="admin", username=None, APIVersion='2011-01-01',
+                 region=None, ec2_ip=None, s3_ip=None, as_ip=None, elb_ip=None, download_creds=True,boto_debug=0,
+                 debug_method=None):
         self.config_file = config_file 
         self.APIVersion = APIVersion
         self.eucapath = "/opt/eucalyptus"
@@ -69,13 +74,14 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
         self.credpath = credpath
         self.download_creds = download_creds
         self.logger = eulogger.Eulogger(identifier="EUCAOPS")
-        self.debug = self.logger.log.debug
+        self.debug = debug_method or self.logger.log.debug
         self.critical = self.logger.log.critical
         self.info = self.logger.log.info
         self.username = username
         self.account_id = None
         self.aws_access_key_id = aws_access_key_id
-        self.aws_secret_access_key = aws_secret_access_key 
+        self.aws_secret_access_key = aws_secret_access_key
+
 
         if self.config_file is not None:
             ## read in the config file
@@ -107,10 +113,11 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
                         self.credpath = self.get_credentials(account,user)
                         self.debug("Successfully downloaded and synced credentials")
                     except Exception, e:
+                        tb = self.get_traceback()
                         self.debug("Caught an exception when getting credentials from first CLC: " + str(e))
                         ### If i only have one clc this is a critical failure, else try on the other clc
                         if len(clc_array) < 2:
-                            raise Exception("Could not get credentials from first CLC and no other to try")
+                            raise Exception(str(tb) + "\nCould not get credentials from first CLC and no other to try")
                         self.swap_clc()
                         self.sftp = self.clc.ssh.connection.open_sftp()
                         self.get_credentials(account,user)
@@ -155,6 +162,15 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
             except Exception, e:
                 self.debug("Unable to create AS connection because of: " + str(e) )
 
+            try:
+                if self.credpath and not elb_ip:
+                    elb_ip = self.get_elb_ip()
+                self.setup_elb_connection(endpoint=elb_ip, path="/services/LoadBalancing", port=8773, is_secure=False, region=region, aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key, boto_debug=boto_debug)
+            except Exception, e:
+                self.debug("Unable to create ELB connection because of: " + str(e) )
+        if self.clc:
+            self.update_property_manager()
+
     def get_available_vms(self, type=None, zone=None):
         """
         Get available VMs of a certain type or return a dictionary with all types and their available vms
@@ -171,20 +187,31 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
                 current_zone = zones[zone_index]
                 if re.search( zone, current_zone.name):
                     break
-                zone_index += 7
-            if zone_index > (len(zones) - 1)   :
-                self.fail("Was not able to find AZ: " + zone)
+                zone_index += 20
+            if zone_index > (len(zones) - 1):
                 raise Exception("Unable to find Availability Zone")    
         else:
             zone = zones[0].name
             
         ### Inline switch statement
-        type_index = {
-                      'm1.small': 2,
-                      'c1.medium': 3,
-                      'm1.large': 4,
-                      'm1.xlarge': 5,
-                      'c1.xlarge': 6,
+        type_index = {  "t1.micro": 2,
+                        "m1.small": 3,
+                        "m1.large": 4,
+                        "m1.xlarge" : 5,
+                        "c1.xlarge" : 6,
+                        "m2.xlarge" : 7,
+                        "c1.medium" : 8,
+                        "m1.medium" : 9,
+                        "m3.xlarge" : 10,
+                        "m2.2xlarge" : 11,
+                        "m3.2xlarge" : 12,
+                        "m2.4xlarge" : 13,
+                        "cc1.4xlarge" : 14,
+                        "hi1.4xlarge" : 15,
+                        "cc2.8xlarge" : 16,
+                        "cg1.4xlarge" : 17,
+                        "cr1.8xlarge" : 18,
+                        "hs1.8xlarge" : 19
                       }[type] 
         type_state = zones[ zone_index + type_index ].state.split()
         self.debug("Finding available VMs: Partition=" + zone +" Type= " + type + " Number=" +  str(int(type_state[0])) )
@@ -206,16 +233,19 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
             raise Exception("Setting property " + property + " failed")
     
    
-    def cleanup_artifacts(self):
+    def cleanup_artifacts(self,instances=True, snapshots=True, volumes=True):
         """
         Description: Attempts to remove artifacts created during and through this eutester's lifespan.
         """
 
         self.debug("Starting cleanup of artifacts")
-        for res in self.test_resources["reservations"]:
-            self.terminate_instances(res)
-        self.clean_up_test_volumes()
-        self.cleanup_test_snapshots()
+        if instances:
+            for res in self.test_resources["reservations"]:
+                self.terminate_instances(res)
+        if volumes:
+            self.clean_up_test_volumes()
+        if snapshots:
+            self.cleanup_test_snapshots()
 
         for key,array in self.test_resources.iteritems():
             for item in array:
@@ -301,13 +331,10 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
         self.debug('clean_up_volumes: Deleteing volumes now...')
         self.delete_volumes(euvolumes)
 
-
-
                     
     def get_current_resources(self,verbose=False):
         '''Return a dictionary with all known resources the system has. Optional pass the verbose=True flag to print this info to the logs
            Included resources are: addresses, images, instances, key_pairs, security_groups, snapshots, volumes, zones
-        
         '''
         current_artifacts = dict()
         current_artifacts["addresses"] = self.ec2.get_all_addresses()
@@ -354,7 +381,7 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
         f = None
         try:
             #f = open(filepath, 'r')
-            self.testconf = EuConfig(filepath)
+            self.testconf = EuConfig(filepath, legacy_qa_config=True)
             f = self.testconf.legacybuf.splitlines()
         except IOError as (errno, strerror):
             self.debug( "ERROR: Could not find config file " + self.config_file)
@@ -373,12 +400,7 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
                 machine_dict["arch"] = machine_details[3]
                 machine_dict["source"] = machine_details[4]
                 machine_dict["components"] = map(str.lower, machine_details[5].strip('[]').split())
-               
-                ### We dont want to login to ESX boxes
-                if re.search("vmware", machine_dict["distro"], re.IGNORECASE):
-                    connect=False
-                else:
-                    connect=True
+
                 ### ADD the machine to the array of machine
                 cloud_machine = Machine(   machine_dict["hostname"], 
                                         distro = machine_dict["distro"], 
@@ -386,7 +408,7 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
                                         arch = machine_dict["arch"], 
                                         source = machine_dict["source"], 
                                         components = machine_dict["components"],
-                                        connect = connect,
+                                        connect = True,
                                         password = self.password,
                                         keypath = self.keypath,
                                         username = username
@@ -403,7 +425,11 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
         #f.close()   
         config_hash["machines"] = machines 
         return config_hash
-    
+
+    def update_property_manager(self,machine=None):
+        machine = machine or self.clc
+        self.property_manager = Euproperty_Manager(self,debugmethod=self.debug)
+
     def swap_clc(self):
         all_clcs = self.get_component_machines("clc")
         if self.clc is all_clcs[0]:
@@ -514,7 +540,7 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops):
         
     
     def download_creds_from_clc(self, admin_cred_dir):
-        self.debug("Downloading credentials from " + self.clc.hostname)
+        self.debug("Downloading credentials from " + self.clc.hostname + ", path:" + admin_cred_dir + "/creds.zip")
         self.sftp.get(admin_cred_dir + "/creds.zip" , admin_cred_dir + "/creds.zip")
         os.system("unzip -o " + admin_cred_dir + "/creds.zip -d " + admin_cred_dir )
